requirements.txt
google_cloud_aiplatform==1.25.0
pandas==2.0.2
pandas==1.3.5
protobuf==3.20.3
python-dotenv==1.0.0
streamlit==1.23.1
streamlit==1.25.0
